requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
numpy==2.2.4
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==2.7.0
transformers==4.50.1
torch==2.6.0
pinecone==6.0.2
